% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psqn-package.R
\docType{package}
\name{psqn-package}
\alias{psqn-package}
\alias{_PACKAGE}
\title{psqn: Partially Separable Quasi-Newton}
\description{
Provides quasi-Newton methods to minimize partially separable
    functions. The methods are largely described by  
    Nocedal and Wright (2006) <doi:10.1007/978-0-387-40065-5>.
}
\details{
The main method in the psqn package is the \code{\link{psqn}} function.
Notice that it is also possible to use the package from C++. This may
yield a large reduction in the computation time. See the vignette for
details e.g. by calling \code{vignette("psqn", package = "psqn")}.
A brief introduction is provided in the "quick-intro" vignette
(see \code{vignette("quick-intro", package = "psqn")}).

This package is fairly new. Thus, results may change and
contributions and feedback is much appreciated.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/boennecd/psqn}
  \item Report bugs at \url{https://github.com/boennecd/psqn/issues}
}

}
\author{
\strong{Maintainer}: Benjamin Christoffersen \email{boennecd@gmail.com}

}
\keyword{internal}
